Split up config among new files for better readability
